backend/requirements.txt
asgiref==3.6.0
Babel==2.10.3
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
distro==1.8.0
Django==3.2.19
django-filter==23.1
django-guardian==2.4.0
djangorestframework==3.14.0
djoser==2.1.0
httplib2==0.20.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
olefile==0.46
Pillow==9.4.0
psycopg2==2.9.5
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
psycopg2-binary==2.9.3
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
wadllib==1.3.6
webcolors==1.11.1
